refactor: remove all JavaScript code
- Converted check-requirements.js to TypeScript
- Removed all .js files from source
- Updated package.json references
- Now 100% TypeScript codebase
